requirements.txt
streamlit-option-menu
pillow
mistralai
google-generativeai
pytesseract
PyPDF2
poppler-utils
PyMuPDF
groq
streamlit-navigation-bar